third_party/blink/web_tests/external/wpt/selection/textcontrols/selectionchange-expected.txt

This is a testharness.js-based test.
Found 12 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Calling select() on the input element
  assert_equals: expected 1 but got 2
[FAIL] Calling select() twice on the input element
  assert_equals: expected 1 but got 2
[FAIL] Calling setRangeText() after select() on the input element
  assert_equals: expected 1 but got 2
[FAIL] Calling setRangeText() repeatedly on the input element
  assert_equals: expected 1 but got 5
[FAIL] Calling setRangeText() after select() on the disconnected input element
  assert_equals: expected 1 but got 2
[FAIL] Calling setRangeText() repeatedly on the disconnected input element
  assert_equals: expected 1 but got 4
[FAIL] Calling select() on the textarea element
  assert_equals: expected 1 but got 2
[FAIL] Calling select() twice on the textarea element
  assert_equals: expected 1 but got 2
[FAIL] Calling setRangeText() after select() on the textarea element
  assert_equals: expected 1 but got 2
[FAIL] Calling setRangeText() repeatedly on the textarea element
  assert_equals: expected 1 but got 5
[FAIL] Calling setRangeText() after select() on the disconnected textarea element
  assert_equals: expected 1 but got 2
[FAIL] Calling setRangeText() repeatedly on the disconnected textarea element
  assert_equals: expected 1 but got 4
Harness: the test ran to completion.
